[+] final push

+++ src/RIPJO/Incident/Incident.cpp
@@ -19,7 +19,7 @@ int RIPJO::Incident::executeIncident(std::mutex &pMutex, std::size_t &influence,
 {
     std::unique_lock<std::mutex> pLock(pMutex);
     std::unique_lock<std::mutex> dLock(dMutex);
-    if (influence - _influenceCost < 0)
+    if ((int)influence - (int)_influenceCost < 0)
         return 1;
     influence -= _influenceCost;
     unrest += _unrestGain;
